requirements.txt
Django==1.7.3
Pillow==2.7.0
django-ckeditor==4.4.7
django-extensions==1.4.9
django-permission==0.8.3
django-suit==0.2.12
django-waffle==0.10.1
factory-boy==2.4.1
sorl-thumbnail==12.2
pytz==2014.10
wsgiref==0.1.2
